+++ packages/synapse-interface/components/StateManagedBridge/OutputContainer.tsx
@@ -65,9 +65,9 @@ export const OutputContainer = ({}) => {
                 `}
                 placeholder="0.0000"
                 value={
-                  bridgeQuote.outputAmountString === '0'
+                  bridgeQuote?.outputAmountString === '0'
                     ? ''
-                    : bridgeQuote.outputAmountString
+                    : bridgeQuote?.outputAmountString
                 }
                 name="inputRow"
                 autoComplete="off"